# src/test/resources/api/openapi.yaml
openapi: 3.0.1
info:
  title: api-checklist
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /checklists/{checklistId}/phases/{phaseId}/tasks:
    get:
      tags:
      - Task resources
      summary: Fetch all tasks in a checklist phase
      operationId: fetchChecklistPhaseTasks
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      - name: phaseId
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
            application/problem+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      tags:
      - Task resources
      summary: Create task in checklist phase
      operationId: createChecklistPhaseTask
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      - name: phaseId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskCreateRequest"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "201":
          description: Successful Operation
          headers:
            Location:
              style: simple
              schema:
                type: string
...
  /employee-checklists/{employeeChecklistId}/tasks/{taskId}:
    patch:
      tags:
      - Employee checklist resources
      summary: Update fulfilment of a task
      operationId: updateTaskFulfilment
      parameters:
      - name: employeeChecklistId
        in: path
        required: true
        schema:
          type: string
      - name: taskId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeChecklistTaskUpdateRequest"
        required: true
      responses:
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmployeeChecklistTask"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/EmployeeChecklistTask"
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /checklists/{checklistId}/phases/{phaseId}/tasks/{taskId}:
    get:
      tags:
      - Task resources
      summary: Fetch task in a checklist phase
      operationId: fetchChecklistPhaseTask
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      - name: phaseId
        in: path
        required: true
        schema:
          type: string
      - name: taskId
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      tags:
      - Task resources
      summary: Delete task in checklist phase
      operationId: deleteChecklistPhaseTask
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      - name: phaseId
        in: path
        required: true
        schema:
          type: string
      - name: taskId
        in: path
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "204":
          description: No Content
    patch:
      tags:
      - Task resources
      summary: Update task in checklist phase
      operationId: updateChecklistPhaseTask
      parameters:
      - name: checklistId
        in: path
        required: true
        schema:
          type: string
      - name: phaseId
        in: path
        required: true
        schema:
          type: string
      - name: taskId
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdateRequest"
        required: true
      responses:
        "404":
          description: Not Found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad Request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal Server Error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "200":
          description: Successful Operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
    StatusType:
      type: object
      properties:
        statusCode:
          type: integer
          format: int32
        reasonPhrase:
          type: string
...
    ConstraintViolationProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        type:
          type: string
          format: uri
        status:
          $ref: "#/components/schemas/StatusType"
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        message:
          type: string
        instance:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    ThrowableProblem:
      type: object
      properties:
        cause:
          $ref: "#/components/schemas/ThrowableProblem"
        stackTrace:
          type: array
          items:
            type: object
            properties:
              classLoaderName:
                type: string
              moduleName:
                type: string
              moduleVersion:
                type: string
              methodName:
                type: string
              fileName:
                type: string
              lineNumber:
                type: integer
                format: int32
              className:
                type: string
              nativeMethod:
                type: boolean
        message:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        parameters:
          type: object
          additionalProperties:
            type: object
        status:
          $ref: "#/components/schemas/StatusType"
        title:
          type: string
        detail:
          type: string
        suppressed:
          type: array
          items:
            type: object
            properties:
              stackTrace:
                type: array
                items:
                  type: object
                  properties:
                    classLoaderName:
                      type: string
                    moduleName:
                      type: string
                    moduleVersion:
                      type: string
                    methodName:
                      type: string
                    fileName:
                      type: string
                    lineNumber:
                      type: integer
                      format: int32
                    className:
                      type: string
                    nativeMethod:
                      type: boolean
              message:
                type: string
              localizedMessage:
                type: string
        localizedMessage:
          type: string
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    QuestionType:
      type: string
      description: The question type of the task
      writeOnly: true
      enum:
      - YES_OR_NO
      - YES_OR_NO_WITH_TEXT
      - COMPLETED_OR_NOT_RELEVANT
      - COMPLETED_OR_NOT_RELEVANT_WITH_TEXT
...
    RoleType:
      type: string
      description: The role type of the task
      readOnly: true
      enum:
      - EMPLOYEE
      - MANAGER
...
    Permission:
      type: string
      description: The permission needed to administrate the phase
      writeOnly: true
      enum:
      - SUPERADMIN
      - ADMIN
...
    TaskCreateRequest:
      required:
      - heading
      - permission
      - questionType
      - roleType
      - sortOrder
      type: object
      properties:
        heading:
          type: string
          description: The name of the task
          writeOnly: true
          example: Name of the task
        text:
          type: string
          description: The body text of the task
          writeOnly: true
          example: Body text of the task
        sortOrder:
          type: integer
          description: The sort order of the task
          format: int32
          writeOnly: true
          example: 1
        roleType:
          $ref: "#/components/schemas/RoleType"
        permission:
          $ref: "#/components/schemas/Permission"
        questionType:
          $ref: "#/components/schemas/QuestionType"
      description: Model for task create request
...
    Task:
      type: object
      properties:
        id:
          type: string
          description: The id of the task
          readOnly: true
          example: 5a6c3e4e-c320-4006-b448-1fd4121df828
        heading:
          type: string
          description: The heading of the task
          readOnly: true
          example: Bjud på fika
        text:
          type: string
          description: The body text of the task
          readOnly: true
          example: Detta är en beskrivning av ett uppdrag
        sortOrder:
          type: integer
          description: The sort order of the task
          format: int32
          readOnly: true
          example: 1
        roleType:
          $ref: "#/components/schemas/RoleType"
        questionType:
          $ref: "#/components/schemas/QuestionType"
        permission:
          $ref: "#/components/schemas/Permission"
        created:
          type: string
          description: The date and time the task was created
          format: date-time
          readOnly: true
          example: 2023-11-22T15:30:00+03:00
        updated:
          type: string
          description: The date and time the task was last updated
          format: date-time
          readOnly: true
          example: 2023-11-22T15:30:00+03:00
      description: Task model
    EmployeeChecklistTaskUpdateRequest:
      type: object
      properties:
        fulfilmentStatus:
          $ref: "#/components/schemas/FulfilmentStatus"
        responseText:
          type: string
          description: The response text for the task fulfilment
          nullable: true
          writeOnly: true
          example: Har bjudit på fika
      description: Model for update request of fulfilment for a task
    FulfilmentStatus:
      type: string
      description: The status of the task fulfilment
      nullable: true
      writeOnly: true
      enum:
      - EMPTY
      - "TRUE"
      - "FALSE"
    EmployeeChecklistTask:
      type: object
      properties:
        id:
          type: string
          description: The id of the task
          readOnly: true
          example: 5a6c3e4e-c320-4006-b448-1fd4121df828
        heading:
          type: string
          description: The heading of the task
          readOnly: true
          example: Bjud på fika
        text:
          type: string
          description: The body text of the task
          readOnly: true
          example: Detta är en beskrivning av ett uppdrag
        sortOrder:
          type: integer
          description: The sort order for the task
          format: int32
          readOnly: true
        roleType:
          $ref: "#/components/schemas/RoleType"
        questionType:
          $ref: "#/components/schemas/QuestionType"
        customTask:
          type: boolean
          description: Tells if the task is only applies to the current checklist
            or not
          readOnly: true
        responseText:
          type: string
          description: The task response text
          readOnly: true
          example: Jag har bjudit på fika
        fulfilmentStatus:
          $ref: "#/components/schemas/FulfilmentStatus"
        updated:
          type: string
          description: The date and time the task was last updated
          format: date-time
          readOnly: true
          example: 2023-11-22T15:30:00+03:00
      description: Model for a employee checklist task
...
    TaskUpdateRequest:
      type: object
      properties:
        heading:
          type: string
          description: The name of the task
          writeOnly: true
        text:
          type: string
          description: The body text of the task
          writeOnly: true
        sortOrder:
          type: integer
          description: The sort order of the task
          format: int32
          writeOnly: true
          example: 1
        roleType:
          $ref: "#/components/schemas/RoleType"
        permission:
          $ref: "#/components/schemas/Permission"
        questionType:
          $ref: "#/components/schemas/QuestionType"
      description: Model for task update request
...
  securitySchemes: {}